handle client errors

+++ app/api/controllers/productCont.js
@@ -1,37 +1,57 @@
 const Product = require('../models/productsDB')
+const validateUser = require('../validations/validateUser')
+const statusObject = require('../helpers/statusObject')
 
-const getAllProduct = async (req, res) => {
+// Public
+const getAllProduct = async (req, res, next) => {
    const productList = await Product.find().populate({
       path: 'addedBy',
       select: 'name email _id',
    })
-   if (!productList) throw new Error('server failure')
    res.status(200).json(productList)
 }
 
-const createProduct = async (req, res) => {
-   const product = new Product(req.body)
+const getProductById = async (req, res, next) => {
+   const product = await Product.findById(req.params.id)
+   if (!product) return next(statusObject(404))
+   res.status(200).json(product)
+}
 
+// Private
+const createProduct = async (req, res, next) => {
+   const product = new Product({ ...req.body, addedBy: req.user._id })
    await product.save()
-   if (!product) throw new Error('invalid data')
+   if (!product) return next(statusObject(400))
    res.status(201).json(product)
 }
-const getProductById = async (req, res) => {
+
+const updateProduct = async (req, res, next) => {
    const product = await Product.findById(req.params.id)
-   if (!product) throw new Error('not found')
-   res.status(200).json(product)
+   if (!product) return next(statusObject(404))
+   if (!validateUser(req.user._id, product.addedBy))
+      return next(statusObject(403))
+
+   const updatedProduct = await Product.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+         new: true,
+         runValidators: true,
+      }
+   )
+   if (!updatedProduct) return next(statusObject(400))
+   res.status(200).json(updatedProduct)
 }
-const updateProduct = async (req, res) => {
-   const product = await Product.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-   })
-   if (!product) throw new Error('not found')
-   res.status(200).json(product)
-}
-const deleteProduct = async (req, res) => {
-   const product = await Product.findByIdAndDelete(req.params.id)
-   if (!product) throw new Error('not found')
-   res.status(204).json(product)
+
+const deleteProduct = async (req, res, next) => {
+   const product = await Product.findById(req.params.id)
+   if (!product) return next(statusObject(404))
+   if (!validateUser(req.user._id, product.addedBy))
+      return next(statusObject(403))
+
+   const deletedProduct = await Product.findByIdAndDelete(req.params.id)
+   if (!deletedProduct) return next(statusObject(500))
+   res.status(204).json(deletedProduct)
 }
 
 module.exports = {
